# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Vivid Detailing API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/services/{id}:
    patch:
      operationId: adminUpdateService
      tags: [admin]
      summary: Update service (price, active state, description)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUpdateServiceBody"
      responses:
        "200":
          description: Updated service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
...
  /admin/add-ons/{id}:
    patch:
      operationId: adminUpdateAddOn
      tags: [admin]
      summary: Update add-on
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUpdateAddOnBody"
      responses:
        "200":
          description: Updated add-on
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AddOn"
  /admin/bookings:
    get:
      operationId: adminListBookings
      tags: [admin]
      summary: List bookings with optional filters
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, completed, cancelled]
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
        - name: serviceCategory
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Bookings list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
  /admin/bookings/{id}:
    patch:
      operationId: adminUpdateBooking
      tags: [admin]
      summary: Update booking (confirm, complete, cancel, condition score, photos)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUpdateBookingBody"
      responses:
        "200":
          description: Updated booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
...
  /admin/customers/{id}:
    patch:
      operationId: adminUpdateCustomer
      tags: [admin]
      summary: Update customer profile (condition score, photos, etc.)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUpdateCustomerBody"
      responses:
        "200":
          description: Updated customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /admin/analytics:
    get:
      operationId: getAnalytics
      tags: [analytics]
      summary: Revenue by category + popular add-ons
      parameters:
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Analytics data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsData"
  /admin/seasonal-promos:
    get:
      operationId: listSeasonalPromos
      tags: [admin]
      summary: List seasonal promos
      responses:
        "200":
          description: Promos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SeasonalPromo"
    patch:
      operationId: updateSeasonalPromo
      tags: [admin]
      summary: Toggle or edit a seasonal promo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSeasonalPromoBody"
      responses:
        "200":
          description: Updated promo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SeasonalPromo"
components:
  schemas:
...
    Service:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        category:
          type: string
        pricingRule:
          type: string
        basePrice:
          type: number
          nullable: true
        isActive:
          type: boolean
        isSeasonal:
          type: boolean
        sortOrder:
          type: integer
        description:
          type: string
          nullable: true
        includes:
          type: array
          items:
            type: string
        showInProtectionStep:
          type: boolean
        prices:
          type: array
          items:
            $ref: "#/components/schemas/ServicePrice"
      required: [id, name, category, pricingRule, isActive, isSeasonal, sortOrder, showInProtectionStep, includes, prices]
    ServicePrice:
      type: object
      properties:
        vehicleType:
          type: string
        price:
          type: number
      required: [vehicleType, price]
    AddOn:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        categoryGroup:
          type: string
        isActive:
          type: boolean
        showInProtectionStep:
          type: boolean
        sortOrder:
          type: integer
        prices:
          type: array
          items:
            $ref: "#/components/schemas/ServicePrice"
      required: [id, name, categoryGroup, isActive, showInProtectionStep, sortOrder, prices]
...
    Booking:
      type: object
      properties:
        id:
          type: string
        customerId:
          type: string
          nullable: true
        vehicleId:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, confirmed, completed, cancelled]
        appointmentAt:
          type: string
          nullable: true
        totalEstimate:
          type: number
          nullable: true
        depositPaid:
          type: boolean
        notes:
          type: string
          nullable: true
        ghlContactId:
          type: string
          nullable: true
        ghlOpportunityId:
          type: string
          nullable: true
        createdAt:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/BookingItem"
        customer:
          $ref: "#/components/schemas/Customer"
          nullable: true
        vehicle:
          $ref: "#/components/schemas/Vehicle"
          nullable: true
      required: [id, status, depositPaid, createdAt, items]
    BookingItem:
      type: object
      properties:
        id:
          type: string
        bookingId:
          type: string
        itemType:
          type: string
          enum: [service, addon, quote, promo]
        itemName:
          type: string
        unitPrice:
          type: number
          nullable: true
        quantity:
          type: integer
        isQuoteBased:
          type: boolean
      required: [id, bookingId, itemType, itemName, quantity, isQuoteBased]
...
    Customer:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        ghlContactId:
          type: string
          nullable: true
        ghlSyncStatus:
          type: string
          nullable: true
        ghlLastSync:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, createdAt]
...
    Vehicle:
      type: object
      properties:
        id:
          type: string
        customerId:
          type: string
        type:
          type: string
          enum: [car, suv, truck, van]
        year:
          type: integer
          nullable: true
        make:
          type: string
          nullable: true
        model:
          type: string
          nullable: true
        colour:
          type: string
          nullable: true
        licensePlate:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, customerId, type, createdAt]
...
    AdminUpdateServiceBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        isActive:
          type: boolean
        isSeasonal:
          type: boolean
        basePrice:
          type: number
        prices:
          type: array
          items:
            $ref: "#/components/schemas/ServicePrice"
        includes:
          type: array
          items:
            type: string
    AdminUpdateAddOnBody:
      type: object
      properties:
        name:
          type: string
        isActive:
          type: boolean
        prices:
          type: array
          items:
            $ref: "#/components/schemas/ServicePrice"
    AdminUpdateBookingBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, confirmed, completed, cancelled]
        appointmentAt:
          type: string
        notes:
          type: string
        conditionScore:
          type: integer
        beforePhotoUrls:
          type: array
          items:
            type: string
        afterPhotoUrls:
          type: array
          items:
            type: string
    AdminUpdateCustomerBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        conditionScore:
          type: integer
    AnalyticsData:
      type: object
      properties:
        revenueByCategory:
          type: array
          items:
            $ref: "#/components/schemas/RevenueStat"
        popularAddOns:
          type: array
          items:
            $ref: "#/components/schemas/AddOnStat"
        totalRevenue:
          type: number
        totalBookings:
          type: integer
        completedBookings:
          type: integer
        pendingBookings:
          type: integer
      required: [revenueByCategory, popularAddOns, totalRevenue, totalBookings, completedBookings, pendingBookings]
    RevenueStat:
      type: object
      properties:
        category:
          type: string
        revenue:
          type: number
        bookingCount:
          type: integer
      required: [category, revenue, bookingCount]
    AddOnStat:
      type: object
      properties:
        name:
          type: string
        count:
          type: integer
        revenue:
          type: number
      required: [name, count, revenue]
    SeasonalPromo:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        basePrice:
          type: number
        isActive:
          type: boolean
        validFrom:
          type: string
          nullable: true
        validTo:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        includes:
          type: array
          items:
            type: string
      required: [id, name, basePrice, isActive, includes]
    UpdateSeasonalPromoBody:
      type: object
      properties:
        id:
          type: string
        isActive:
          type: boolean
        basePrice:
          type: number
        name:
          type: string
        description:
          type: string
      required: [id]